Push initialization code into base class.
